cxx-qt-gen: move headers to cxx-qt crate
