+++ crates/cxx-qt-build/Cargo.toml
@@ -16,6 +16,7 @@ repository.workspace = true
 cc.workspace = true
 convert_case.workspace = true
 cxx-gen.workspace = true
+cxx-qt.workspace = true
 cxx-qt-lib-headers.workspace = true
 cxx-qt-gen.workspace = true
 proc-macro2.workspace = true
