+++ crates/cxx-qt/include/cxxqt_signalhandler.h
